+++ sumologic/provider.go
@@ -116,6 +116,7 @@ func Provider() terraform.ResourceProvider {
 			"sumologic_log_search":                               resourceSumologicLogSearch(),
 			"sumologic_metrics_search":                           resourceSumologicMetricsSearch(),
 			"sumologic_rum_source":                               resourceSumologicRumSource(),
+			"sumologic_azure_event_hub_log_source":               resourceSumologicGenericPollingSource(),
 		},
 		DataSourcesMap: map[string]*schema.Resource{
 			"sumologic_cse_log_mapping_vendor_product": dataSourceCSELogMappingVendorAndProduct(),